requirements.txt
absl-py==1.4.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
configparser==6.0.0
contourpy==1.1.0
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.42.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.7.2
mediapipe==0.10.3
numpy==1.25.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
protobuf==3.20.3
pycparser==2.21
PyGLM==2.7.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.0.9
PySDL2==0.9.16
pysdl2-dll==2.28.2
python-dateutil==2.8.2
six==1.16.0
sounddevice==0.4.6
urllib3==2.0.4
